pyproject.toml
[project]
dynamic = ["version"]
name = "bot_offer_news"

requires-python = ">= 3.12"

dependencies = [
    "aiogram == 3.8.0",
    "asyncpg==0.29.0",
    "psycopg-binary==3.1.19",
    "SQLAlchemy == 2.0.31",
    "alembic==1.13.1",
    "python-dotenv==1.0.1",
    "dishka==1.2.0",
    "fastapi==0.111.0"
]